# backend/api_v2/swagger.yaml
openapi: 3.0.3
info:
  title: Oto API
  description: |
    The Oto API is a FastAPI-based backend service for conversation analysis. 
    It processes audio files, transcribes them, and provides AI-powered analysis 
    including summaries, highlights, insights, and sentiment analysis.
  version: 1.0.0
  contact:
    name: Oto API Support
servers:
  - url: http://localhost:8000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Privy access token
  parameters:
    UserIdHeader:
      name: Oto-User-Id
      in: header
      required: true
      schema:
        type: string
      description: User ID for authentication
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        name:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        nationality:
          type: string
          nullable: true
        first_language:
          type: string
          nullable: true
        second_languages:
          type: string
          nullable: true
        interests:
          type: string
          nullable: true
        preferred_topics:
          type: string
          nullable: true
    UpdateUser:
      type: object
      properties:
        name:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        nationality:
          type: string
          nullable: true
        first_language:
          type: string
          nullable: true
        second_languages:
          type: string
          nullable: true
        interests:
          type: string
          nullable: true
        preferred_topics:
          type: string
          nullable: true
...
    UserCreateResponse:
      type: object
      properties:
        id:
          type: string
          description: Created user ID
      required:
        - id
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
paths:
...
  /user/create:
    post:
      tags:
        - User Management
      summary: Create a new user account
      description: Creates a new user account with the provided user ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/get:
    get:
      tags:
        - User Management
      summary: Get user profile information
      description: Retrieves the current user's profile information
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/update:
    post:
      tags:
        - User Management
      summary: Update user profile information
      description: Updates the current user's profile with provided information
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...